refactor(upload > components > watermark button with loader): extract from WatermarkSettingsView
